Changed randomization
Randomization now fully shuffles the list of statements instead of randomizing just the statement string.
Updated step3.html accordingly:
Now we filter for uncategorized statements, picking 1 item at time with limitTo .
Got rid of ng-class=... at  <li> element because it was unused and finally, got rid of ng-hide at <li> because it is not needed anymore.